Issue #249 - extracted http class out of testing code
